StreamCreator.cs, Connection.cs, Exception.cs, Logger.cs, MySqlConnectionStringBuilder.cs MySqlPool.cs, MySqlStream.cs, ProcedureCache.cs, SchemaProvider.cs, MySqlByte.cs MySqlUInt32.cs, ------------------------ Compact Framework compatibility CHANGES
datareader.cs - Moved literal string to resource

+++ Driver/Source/MySqlPool.cs
@@ -51,15 +51,15 @@ namespace MySql.Data.MySqlClient
             inUsePool = new List<Driver>((int)maxSize);
             idlePool = new Queue<Driver>((int)maxSize);
 #else
-			inUsePool =new ArrayList(maxSize);
-			idlePool = new Queue(maxSize);
+			inUsePool =new ArrayList((int)maxSize);
+			idlePool = new Queue((int)maxSize);
 #endif
 
 			// prepopulate the idle pool to minSize
 			for (int i=0; i < minSize; i++) 
 				CreateNewPooledConnection();
 
-            procedureCache = new ProcedureCache(settings.ProcedureCacheSize);
+            procedureCache = new ProcedureCache((int)settings.ProcedureCacheSize);
 		}
 
 		public MySqlConnectionStringBuilder	Settings
